Switches to using `bigint` for fixed-length values, rather than Bytes.
Working with `Uint8Array` is more challenging than working with `bigint`, and it is standard unlike `Bytes`.  This means we can drop the dependency on `@zoltu/ethereum-types` (yay for less dependencies!) and still provide the user with a pretty good UX.  It also simplifies decoding/encoding a tad (though not much).